+++ test/testenv.jl
@@ -37,6 +37,15 @@ if !@isdefined(testenv_defined)
 
     function addprocs_with_testenv(X; rr_allowed=true, kwargs...)
         exename = rr_allowed ? `$rr_exename $test_exename` : test_exename
+        if X isa Integer
+            if Sys.iswindows()
+                heap_size=round(Int,(Sys.free_memory()/(1024^2)/(X+1)))
+                heap_size -= 300 # I don't know anymore
+            else
+                heap_size=round(Int,(Sys.total_memory()/(1024^2)/(X+1)))
+            end
+            push!(test_exeflags.exec, "--heap-size-hint=$(heap_size)M")
+        end
         addprocs(X; exename=exename, exeflags=test_exeflags, kwargs...)
     end